fjjf

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -11,7 +11,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	for (int i = 0; i < size; i++)
+	for (size_t i = 0; i < size; i++)
 	{
 		printf("%d \n", *array);
 		array++;
